Implemented EthernetTCP, a somewhat non-blocking version of EthernetClient

+++ src/EthernetClient.cpp
@@ -50,7 +50,7 @@ int EthernetClient::connect(IPAddress ip, uint16_t port)
 #if defined(ESP8266) || defined(ESP32)
 	if (ip == IPAddress((uint32_t)0) || ip == IPAddress(0xFFFFFFFFul)) return 0;
 #else
-	if (ip == IPAddress(0ul) || ip == IPAddress(0xFFFFFFFFul)) return 0;
+	if (ip == IPAddress((uint32_t)0ul) || ip == IPAddress(0xFFFFFFFFul)) return 0;
 #endif
 	_sockindex = Ethernet.socketBegin(SnMR::TCP, 0);
 	if (_sockindex >= MAX_SOCK_NUM) return 0;
